+++ CHANGELOG.md
@@ -7,11 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Changed
+- Change style of checkbox in the admin page.
+
 ### Fixed
 - Fix the links social media in the footer.
 - Fix margin in about info.
 - Fix shortcut search in mac.
 - Fix remove image preview when the user remove the image in the admin page.
+- Fix checkbox in view page.
 
 ## [1.5.0] - 06/06/2024